Configurable number of arenas
