+++ options/customizable_test.cc
@@ -1982,8 +1982,10 @@ TEST_F(LoadCustomizableTest, LoadMemoryAllocatorTest) {
   } else {
     ASSERT_NOK(s);
     for (const auto& failure : failures) {
-      if (failure == JemallocNodumpAllocator::kClassName()) {
-        ASSERT_FALSE(JemallocNodumpAllocator::IsSupported());
+      if (failure ==
+          JemallocNodumpAllocator<0 /* kLog2NumArenas */>::kClassName()) {
+        ASSERT_FALSE(
+            JemallocNodumpAllocator<0 /* kLog2NumArenas */>::IsSupported());
       } else if (failure == MemkindKmemAllocator::kClassName()) {
         ASSERT_FALSE(MemkindKmemAllocator::IsSupported());
       } else {
